Subject component now displays more info and global theme setup
I wanted to redo the subject component so that it will display the meaning as well as readings if applicable. This is to replicate what wanikani themselves do. 

Switched to a sorta janky (?) global theme that will later allow for support for dark themes and centralizes the various colors so I don't have to change them in a bunch of places.